issues-5: prevent additional sub package 'url' by using vanityurl as package name
