+++ README.md
@@ -67,9 +67,9 @@ You can browse available artifacts through [Magnolia's Nexus](https://nexus.magn
 You can configure the following settings in the module configuration by JCR or yaml:
 * _excludes_ : Pattern of urls, which are no candidates for vanity urls.
   * by default an exclude for all urls containing a dot is configured, that prevents the virtual uri mapping checks every ordinary request like script.js or page.html 
-* _publicUrlService_ : Implementation of _de.ibmix.magkit.vanity.url.PublicUrlService_. Two implementations are already available.
-  * _de.ibmix.magkit.vanity.url.DefaultPublicUrlService_ (default) : Use of default base url and site configuration with context path replacement.
-  * _de.ibmix.magkit.vanity.url.SimplePublicUrlService_ : Uses configured public prefix and removes the author context path.
+* _publicUrlService_ : Implementation of _de.ibmix.magkit.vanityurl.PublicUrlService_. Two implementations are already available.
+  * _de.ibmix.magkit.vanityurl.DefaultPublicUrlService_ (default) : Use of default base url and site configuration with context path replacement.
+  * _de.ibmix.magkit.vanityurl.SimplePublicUrlService_ : Uses configured public prefix and removes the author context path.
 * _headlessEndpoint_ : is used for headless support and defines the pages rest endpoint which should be checked for vanity urls
 
 #### Headless support (with version 1.6.1)
@@ -78,7 +78,7 @@ For using the headless support you have to decorate the default vanity url virtu
 if you want to support both renderings. All you need is the following yaml which you can place in your module as decoration of the vanity url module 
 (your-module/decorations/magkit-vanity-url/virtualUriMappings) or as own virtual uri mapping (your-module/virtualUriMappings).
 ```yaml
-class: de.ibmix.magkit.vanity.url.HeadlessVirtualVanityUriMapping
+class: de.ibmix.magkit.vanityurl.HeadlessVirtualVanityUriMapping
 ```
 
 ## License
