+++ src/main/java/de/ibmix/magkit/vanityurl/app/PreviewAction.java
@@ -1,4 +1,4 @@
-package de.ibmix.magkit.vanity.url.app;
+package de.ibmix.magkit.vanityurl.app;
 
 /*
  * #%L
@@ -20,7 +20,7 @@ package de.ibmix.magkit.vanity.url.app;
  * #L%
  */
 
-import de.ibmix.magkit.vanity.url.VanityUrlService;
+import de.ibmix.magkit.vanityurl.VanityUrlService;
 import info.magnolia.ui.ValueContext;
 import info.magnolia.ui.api.action.AbstractAction;
 import info.magnolia.ui.api.location.DefaultLocation;
@@ -34,23 +34,22 @@ import javax.inject.Inject;
 import javax.jcr.Node;
 
 import static info.magnolia.ui.api.location.Location.LOCATION_TYPE_APP;
-import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 
 /**
- * Action to see the big QR code image for a vanity url.
+ * Preview action for vanity urls. Opens the website page or the external url in the configured app.
  *
  * @author frank.sommer
- * @since 28.05.14
+ * @since 06.05.14
  */
-public class QrCodeViewAction extends AbstractAction<OpenLocationActionDefinition> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(QrCodeViewAction.class);
+public class PreviewAction extends AbstractAction<OpenLocationActionDefinition> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PreviewAction.class);
 
     private final LocationController _locationController;
     private final ValueContext<Node> _valueContext;
     private VanityUrlService _vanityUrlService;
 
     @Inject
-    public QrCodeViewAction(OpenLocationActionDefinition definition, ValueContext<Node> valueContext, LocationController locationController) {
+    public PreviewAction(OpenLocationActionDefinition definition, ValueContext<Node> valueContext, LocationController locationController) {
         super(definition);
         _locationController = locationController;
         _valueContext = valueContext;
@@ -58,13 +57,11 @@ public class QrCodeViewAction extends AbstractAction<OpenLocationActionDefinitio
 
     @Override
     public void execute() {
-        LOGGER.debug("Execute qr code view action ...");
+        LOGGER.debug("Execute preview action ...");
         if (_valueContext.getSingle().isPresent()) {
-            String link = _vanityUrlService.createImageLink(_valueContext.getSingle().get());
-            if (isNotEmpty(link)) {
-                Location location = new DefaultLocation(LOCATION_TYPE_APP, getDefinition().getAppName(), getDefinition().getSubAppId(), link);
-                _locationController.goTo(location);
-            }
+            String link = _vanityUrlService.createPreviewUrl(_valueContext.getSingle().get());
+            Location location = new DefaultLocation(LOCATION_TYPE_APP, getDefinition().getAppName(), getDefinition().getSubAppId(), link);
+            _locationController.goTo(location);
         }
     }
