+++ src/main/resources/mgnl-bootstrap/magkit-vanity-url/config.modules.magnolia-vanity-url.config.publicUrlService.xml
@@ -7,7 +7,7 @@
     <sv:value>d7bef2a5-61e4-4d14-a7a6-bbab5e2a135e</sv:value>
   </sv:property>
   <sv:property sv:name="class" sv:type="String">
-    <sv:value>de.ibmix.magkit.vanity.url.DefaultPublicUrlService</sv:value>
+    <sv:value>de.ibmix.magkit.vanityurl.DefaultPublicUrlService</sv:value>
   </sv:property>
   <sv:property sv:name="jcr:createdBy" sv:type="String">
     <sv:value>admin</sv:value>
